// src/Utils/enums.ts
export enum productCategory {
  beauty = "beauty",
  fragrances = "fragrances",
  skinCare = "skin-care",
  furniture = "furniture",
  groceries = "groceries",
  homeDecoration = "home-decoration",
  kitchenAccessories = "kitchen-accessories",
  laptops = "laptops",
  smartphones = "smartphones",
  mobileAccessories = "mobile-accessories",
  tablets = "tablets",
  sunglasses = "sunglasses",
  mensShirts = "mens-shirts",
  mensShoes = "mens-shoes",
  mensWatches = "mens-watches",
  tops = "tops",
  womensDresses = "womens-dresses",
  womensShoes = "womens-shoes",
  womensWatches = "womens-watches",
  womensBags = "womens-bags",
  womensJewellery = "womens-jewellery",
  motorcycle = "motorcycle",
  vehicle = "vehicle",
  sportsAccessories = "sports-accessories",
}